layers: fix how SCTP chunk selector behaves in Strict mode
Previously, Strict mode applied to the NextLayerType() function but I
chose to return an error from DecodeFromBytes() instead.

This decision highlights the purpose of this type: it should trivially
enable DecodingLayerParser to process all chunks of an SCTP packet. The
Strict mode was only ever added to fail early when encountered with
unexpected chunk types.