defrag/sctpdefrag: resepect DATA chunks with UserData inside the layer
This commit appears strange, because it is. I've been working two paths
at the same time: implementing the DecodingLayer API for SCTP chunks (in
the gopacket/layers package), and defragmenting messages over SCTP (in
package gopacket/defrag/sctpdefrag).

In my journey to understand the decisions made by maintainers over time,
I've come to several conclusions. Most notably, the UserData field of
DATA chunks must be part of the chunk layer, not a standalone Payload
layer. See the commit, that reverts this decision by past maintainers,
for more details.

This commit then reworks the defragmentation mechanism to respect this
decision, though the history of this branch may appear broken now.