go.mod: upgrade minimum Go version from 1.21 to 1.24
After many attempts to avoid incrementing the minimum Go version for
this module, I've found it too tedious to avoid it.

Though being true to the facts, the benefits present themselves only in
test code, so far, so this may be reverted without impacting the
exported API I will extend in my endevours.