gopacket: register application-specific layer types without collisions
When application-specific packages, like defragmenting SCTP messages,
need to register a layer, it is an effort of coordination between all
the packages in a compiled executable. Unfortunately, this is not always
possible since not all packages can know about each other in advance.

The new function registers such layer types using the first available
index. This means that the number identifying the layer may change
between different users of the same package that registers the layer.
Users are expected to pay attention to this fine print.