defrag/defragtest: expose fragmentDataSource as FragmentSource
Exposing this type allows callers to pick and choose how to read packet
data. For example, tailor-made retransmission patterns are easy to
compose per test by calling ReadFragmentData with the appropriate
positions.

Along the way I've improved test names to better reflect the current
state of the package, as well as extended the coverage to the newly
exported type and its methods.